requirements.txt
psycopg2-binary==2.9.10
pyTelegramBotAPI==4.27.0